requirements.txt
discord.py==2.2.2
python-dateutil==2.8.2
wit==6.0.1
